requirements.txt
boto3>=1.28.0
pandas>=1.5.0
pyspark>=3.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
